Rename summary to aggregate
