+++ test/compile/data/aggregate.test.ts
@@ -3,7 +3,7 @@
 import {assert} from 'chai';
 
 import {DataComponent} from '../../../src/compile/data/data';
-import {summary} from '../../../src/compile/data/summary';
+import {summary} from '../../../src/compile/data/aggregate';
 import {parseUnitModel} from '../../util';
 
 describe('compile/data/summary', function () {
